refactor: rename `ChildTable` -> `Grid`
